Removed BattlePluginsAPI
People have been reporting this impossible error as a BattleArena bug.

Yes, it was impossible for BattleArena to produce this error.

Finally, someone produced a FULL server.log including the error and I was
able to trace it to BattleTracker.

Finally fixed. Another one bite the dust!

Good job team!