Made signs easier to see
https://github.com/BattlePlugins/BattleTracker/issues/8

The new Minecraft colors for &6-gold, &a-neon_green, &e-yellow are harder
to see in 1.8, so they were changed.

Also, there was a problem with light-gray being displayed by default
(instead of black by default). So anything that previously had no color
speicfied, is now &0 - black.

Ideally, this should all be configurable tho.